prjoxide: update clap
Clap was fixed to `<=3.0.0-beta.2` in #12. Upon further investigation,
`beta.4` was functioning as intended; it simply bumped clap's MSRV to
1.54 (see clap-rs/clap#2741).

Clap 3.1 supports Rust as early as 1.54. Upgrade to that (with a few
fixes for a rename that happened in `3.0.0-beta.5`), since it's been out
for well over a year now.